Add support for perl 5.10 and 5.8
The main problem solved here is that older perls haven't installed
the method GLOBREF at the time they call MODIFY_CODE_ATTRIBUTES
on the coderef, so the name cna't be known until afterward.  This
requires some kind of delayed hook, and needs to happen before the
first time the user attempts to use the %EXPORT hash.