Change generator prefix to _generateREFTYPE
The _generateScalar and etc choice of generator name seemed unfortunate,
since the refs are normally uppercase.  I reworked the related code
so that both camel case and uppercase are detected.  Also found some
weird behavior when giving multiple names in an Export() list, and
cleaned that up so that all names get added to all tags, which I think
is more expected behavior than only the first name.