New internal representation for generators
There was a fundamental oversight in choosing REF-ref as the indication
that a field of %EXPORT was a generator rather than a literal export,
because a scalar export could itself be a REF-ref if the package scalar
happened to point to any ref type, which is a fairly common thing for
it to point to.

I reversed the notation for generators so that they are simply scalars
to reference a method name, or method coderef for any non-sub export,
or a coderef-ref for sub exports.

  '$foo' => $generator_method_name,
  'foo'  => $generator_method_name,
  '$foo' => \&generator_method,
  'foo'  => \\&generator_method,

This change does not change the public API, but it does break anything
that was relying on the internal representation.