Fix unit tests for method attributes
I had commented out the unit tests for options with parameters
because is_deeply can't compare anonymous coderefs.  I changed
the mechanism of the test to eval the entire is_deeply line
instead of trying to eval individual pieces, because the latter
was very confusing to read.  Now that it's evaled code, I can
just test that the options with parameters created coderefs.

Meanwhile, I also discovered that the test for EXPORT_TAGS was
not getting run at all.  I fixed that, and fixed some typos in
the code.

I also added a test for option-with-params notation '-(?)'
as reported in #2.